--- frontend/src/Profile/index.tsx
@@ -1,3 +1,0 @@
-import { Profile } from './components/profile'
-
-export { Profile }
