+++ frontend/src/User/index.tsx
@@ -1,0 +1,3 @@
+import { Profile } from './components/Profile/profile'
+
+export { Profile }
